Refactor data initialization to use DATA_INIT_MODE variable and consolidate scripts (fix code duplication)
