+++ src/V_APP/.env.example
@@ -103,4 +103,4 @@ LOKI_URL=10.255.32.132
 # --- Data init ---
 # demo  = full dataset (historical data, many appointments)
 # trial = minimal seed (a truck with a delivery in_transit for live demo flow)
-DATA_INIT_SCRIPT=demo
+DATA_INIT_MODE=trial
