+++ src/V_APP/Data_Module/entrypoint.sh
@@ -47,8 +47,8 @@ else
 fi
 
 # Step 3: Seed data (trigger fires on INSERT, arrival_id is generated)
-echo "Running database initialization (DATA_INIT_SCRIPT=${DATA_INIT_SCRIPT:-demo})..."
-python scripts/data_init_${DATA_INIT_SCRIPT:-demo}.py
+echo "Running database initialization (DATA_INIT_MODE=${DATA_INIT_MODE:-demo})..."
+python scripts/data_init_base.py --mode ${DATA_INIT_MODE:-demo}
 if [[ $? -eq 0 ]]; then
   echo "Data initialization completed successfully"
 else
