+++ src/V_APP/docker-compose.yml
@@ -149,7 +149,7 @@ services:
       - KC_ADMIN_PASSWORD=${KC_ADMIN_PASSWORD}
       - DEFAULT_WORKER_PASSWORD=${DEFAULT_WORKER_PASSWORD}
       - DEFAULT_DRIVER_PASSWORD=${DEFAULT_DRIVER_PASSWORD}
-      - DATA_INIT_SCRIPT=${DATA_INIT_SCRIPT:-demo}
+      - DATA_INIT_MODE=${DATA_INIT_MODE:-demo}
     ports:
       - "8080:8000"
     healthcheck:
